hybris: egl: Move most of wayland window code to common platform
